refactor: change folders/files names
